[pulsewrap] Perform safer PID logging synchronization
The redirection operator was previously actually outside of the
synchronized call, which means the actual write was not synchronized at
all. The script would synchronize the PID echo, but not the write
itself.

To fix this, wrap the echo and the write into a function and synchronize
calls to that function instead.