[backup.sh] Encrypt the backups
The script asks for a passphrase and attempts to decrypt a
passphrase_check file in the backup directory. If it decrypts
successfully, it proceeds with the backup.

The passphrase is given as a herestring to FD 3. The shebang has been
changed to a bash shebang because herestrings are a bashism.